requirements.txt
langchain==0.3.19
langchain-openai==0.3.7
langchain-community==0.3.18
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.2
click==8.1.8
aiohttp==3.11.13
multidict==6.1.0
urllib3==2.3.0
pyyaml==6.0.2
bs4==0.0.2
beautifulsoup4==4.13.3
fastapi==0.115.10
sqlalchemy==2.0.38
pydantic==2.10.6
starlette==0.46.0
orjson==3.10.15
requests==2.32.3
websockets==15.0
anyio==4.8.0
typer==0.15.2
jinja2==3.1.5
httpx==0.28.1
pydantic-settings==2.8.1
markupsafe==3.0.2
uvicorn==0.34.0
idna==3.10
charset-normalizer==3.4.1
chardet==5.2.0
cryptography==44.0.1
openssl==3.0.15
certifi==2025.1.31
setuptools==75.8.0
pygments==2.19.1
python-dotenv==1.0.1
langchain-core==0.3.40
streamlit==1.42.2